Create a Traveler page
Plus an option in the registration when you choose that you are a Traveler so when logging in it will send you to the Traveler page. Plus I added things like looking at the password and hiding it, a line that shows the strength of the password and if it is strong or weak.